충돌 후 수정
