+++ src/main/java/org/example/mobble/board/service/BoardService.java
@@ -11,8 +11,8 @@ import org.example.mobble.board.domain.BoardRepository;
 import org.example.mobble.board.domain.SearchOrderCase;
 import org.example.mobble.board.dto.BoardRequest;
 import org.example.mobble.board.dto.BoardResponse;
-import org.example.mobble.category.Category;
-import org.example.mobble.category.CategoryRepository;
+import org.example.mobble.category.domain.Category;
+import org.example.mobble.category.domain.CategoryRepository;
 import org.example.mobble.report.domain.Report;
 import org.example.mobble.report.domain.ReportCase;
 import org.example.mobble.report.domain.ReportRepository;
