+++ src/main/java/org/example/mobble/category/service/CategoryService.java
@@ -1,11 +1,14 @@
 package org.example.mobble.category.service;
 
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import org.example.mobble._util.error.ex.Exception400;
 import org.example.mobble._util.error.ex.Exception401;
 import org.example.mobble._util.error.ex.Exception404;
 import org.example.mobble.category.domain.Category;
 import org.example.mobble.category.domain.CategoryRepository;
+import org.example.mobble.user.domain.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,8 +20,11 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
 
+    // 연관관계 주입을 위해 User 레퍼런스를 가져오기 위함
+    @PersistenceContext
+    private EntityManager em;
 
-     // 특정 유저의 카테고리 목록 조회
+    // 특정 유저의 카테고리 목록 조회
     @Transactional(readOnly = true)
     public List<Category> getCategoriesByUser(Integer userId) {
         if (userId == null) throw new Exception400("userId가 없습니다.");
@@ -32,13 +38,21 @@ public class CategoryService {
         if (categoryName == null || categoryName.trim().isEmpty()) {
             throw new Exception400("카테고리명이 비어 있습니다.");
         }
-        if (categoryRepository.existsByUserIdAndCategory(userId, categoryName.trim())) {
+        String normalized = categoryName.trim();
+
+        if (categoryRepository.existsByUserIdAndCategory(userId, normalized)) {
             throw new Exception400("이미 존재하는 카테고리명입니다.");
         }
+
+        // 연관관계 주입: userId -> User 엔티티
+        User user = em.find(User.class, userId);
+        if (user == null) throw new Exception404("사용자를 찾을 수 없습니다.");
+
         Category category = Category.builder()
-                .userId(userId)
-                .category(categoryName.trim())
+                .user(user)
+                .category(normalized)
                 .build();
+
         return categoryRepository.save(category);
     }
 
@@ -55,8 +69,8 @@ public class CategoryService {
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new Exception404("카테고리를 찾을 수 없습니다."));
 
-        // 소유자 검증
-        if (!category.getUserId().equals(userId)) {
+        // 소유자 검증: userId -> category.user.id
+        if (category.getUser() == null || !category.getUser().getId().equals(userId)) {
             throw new Exception401("본인 카테고리만 수정할 수 있습니다.");
         }
 
@@ -65,7 +79,8 @@ public class CategoryService {
             throw new Exception400("이미 존재하는 카테고리명입니다.");
         }
 
-        category.rename(normalized); // 더티체킹 반영
+        // 도메인 메서드가 없으므로 세터 사용 (@Data)
+        category.setCategory(normalized); // 더티체킹 반영
         return category;
     }
 
@@ -79,10 +94,11 @@ public class CategoryService {
                 .orElseThrow(() -> new Exception404("카테고리를 찾을 수 없습니다."));
 
         // 소유자 검증
-        if (!category.getUserId().equals(userId)) {
+        if (category.getUser() == null || !category.getUser().getId().equals(userId)) {
             throw new Exception401("본인 카테고리만 삭제할 수 있습니다.");
         }
 
+        // (다음 단계에서: board_tb.category_id = NULL 처리 추가 예정)
         categoryRepository.deleteById(category.getId());
     }
 }
